Refactor and implement send_to_node() method
