--- p2panda-tests/src/test_data/mod.rs
@@ -1,1 +1,0 @@
-pub mod utils;
